Restructure the form folder
