+++ components/form/components/Education.jsx
@@ -1,6 +1,6 @@
 import FormButton from "./FormButton";
 import React, { useContext } from "react";
-import { ResumeContext } from "../../pages/builder";
+import {ResumeContext} from "../../../pages/builder";
 
 const Education = () => {
     const { resumeData, setResumeData} = useContext(ResumeContext);
@@ -10,7 +10,7 @@ const Education = () => {
       newEducation[index][e.target.name] = e.target.value;
       setResumeData({ ...resumeData, education: newEducation });
     };
-  
+
     const addEducation = () => {
       setResumeData({
         ...resumeData,
@@ -20,14 +20,14 @@ const Education = () => {
         ],
       });
     };
-  
+
     const removeEducation = (index) => {
       const newEducation = [...resumeData.education];
       newEducation[index] = newEducation[newEducation.length - 1];
       newEducation.pop();
       setResumeData({ ...resumeData, education: newEducation });
     };
-    
+
     return (
       <div className="flex-col-gap-2">
         <h2 className="input-title">Education</h2>
