+++ components/form/components/LoadUnload.jsx
@@ -1,6 +1,6 @@
 import { FaCloudUploadAlt, FaCloudDownloadAlt } from "react-icons/fa";
 import React, { useContext } from "react";
-import { ResumeContext } from "../../pages/builder";
+import {ResumeContext} from "../../../pages/builder";
 
 const LoadUnload = () => {
   const { resumeData, setResumeData } = useContext(ResumeContext);
