+++ components/form/components/Skill.jsx
@@ -1,5 +1,5 @@
 import React, { useContext } from "react";
-import { ResumeContext } from "../../pages/builder";
+import {ResumeContext} from "../../../pages/builder";
 import FormButton from "./FormButton";
 
 const Skill = ({ title }) => {
