+++ components/form/components/SocialMedia.jsx
@@ -1,6 +1,6 @@
 import FormButton from "./FormButton";
 import React, { useContext } from "react";
-import { ResumeContext } from "../../pages/builder";
+import {ResumeContext} from "../../../pages/builder";
 
 const SocialMedia = () => {
   const { resumeData, setResumeData } = useContext(ResumeContext);
